+++ src/dto/redis/community-event.dto.ts
@@ -1,9 +1,9 @@
-export class CommunityEventDto {
-  eventId: string;
-  eventName: string;
-  organizerId: string;
-  voiceChannelId: string;
-  guildId: string;
+export class RedisCommunityEventDto {
+  id: string;
+  title: string;
+  organizerSId: string;
+  voiceChannelSId: string;
+  guildSId: string;
   startDate: string;
   endDate: string;
   isActive?: boolean;